@@ -15,10 +15,18 @@
 import argparse
 import collections.abc
 import itertools
-from typing import Any, Iterable, Iterator, List, Mapping, Set, Tuple, Union
+from collections import OrderedDict
+from typing import Any, Iterable, Iterator, List, Mapping, Sequence, Set, Tuple, Union
 
 from nox._decorators import Call, Func
 from nox.sessions import Session, SessionRunner
+
+WARN_PYTHONS_IGNORED = "python_ignored"
+
+
+def _unique_list(*args: str) -> List[str]:
+    """Return a list without duplicates, while preserving order."""
+    return list(OrderedDict.fromkeys(args))
 
 
 class Manifest:
@@ -132,6 +140,16 @@ class Manifest:
         if missing_sessions:
             raise KeyError("Sessions not found: {}".format(", ".join(missing_sessions)))
 
+    def filter_by_python_interpreter(self, specified_pythons: Sequence[str]) -> None:
+        """Filter sessions in the queue based on the user-specified
+        python interpreter versions.
+
+        Args:
+            specified_pythons (Sequence[str]): A list of specified
+                python interpreter versions.
+        """
+        self._queue = [x for x in self._queue if x.func.python in specified_pythons]
+
     def filter_by_keywords(self, keywords: str) -> None:
         """Filter sessions using pytest-like keyword expressions.
 
@@ -159,6 +177,33 @@ class Manifest:
                 bound to this manifest and configuration.
         """
         sessions = []
+
+        # if backend is none we wont parametrize the pythons
+        backend = (
+            self._config.force_venv_backend
+            or func.venv_backend
+            or self._config.default_venv_backend
+        )
+        if backend == "none" and isinstance(func.python, (list, tuple, set)):
+            # we can not log a warning here since the session is maybe deselected.
+            # instead let's set a flag, to warn later when session is actually run.
+            func.should_warn[WARN_PYTHONS_IGNORED] = func.python
+            func.python = False
+
+        if self._config.extra_pythons:
+            # If extra python is provided, expand the func.python list to
+            # include additional python interpreters
+            extra_pythons = self._config.extra_pythons  # type: List[str]
+            if isinstance(func.python, (list, tuple, set)):
+                func.python = _unique_list(*func.python, *extra_pythons)
+            elif not multi and func.python:
+                # If this is multi, but there is only a single interpreter, it
+                # is the reentrant case. The extra_python interpreter shouldn't
+                # be added in that case. If func.python is False, the session
+                # has no backend; if None, it uses the same interpreter as Nox.
+                # Otherwise, add the extra specified python.
+                assert isinstance(func.python, str)
+                func.python = _unique_list(func.python, *extra_pythons)
 
         # If the func has the python attribute set to a list, we'll need
         # to expand them.
