@@ -28,6 +28,7 @@ from typing import (
     Mapping,
     Optional,
     Sequence,
+    Tuple,
     Union,
 )
 
@@ -36,7 +37,7 @@ import py
 from nox import _typing
 from nox._decorators import Func
 from nox.logger import logger
-from nox.virtualenv import CondaEnv, ProcessEnv, VirtualEnv
+from nox.virtualenv import CondaEnv, PassthroughEnv, ProcessEnv, VirtualEnv
 
 if _typing.TYPE_CHECKING:
     from nox.manifest import Manifest
@@ -67,6 +68,36 @@ def _normalize_path(envdir: str, path: Union[str, bytes]) -> str:
             )
 
     return full_path
+
+
+def _dblquote_pkg_install_args(args: Tuple[str, ...]) -> Tuple[str, ...]:
+    """Double-quote package install arguments in case they contain '>' or '<' symbols"""
+
+    # routine used to handle a single arg
+    def _dblquote_pkg_install_arg(pkg_req_str: str) -> str:
+        # sanity check: we need an even number of double-quotes
+        if pkg_req_str.count('"') % 2 != 0:
+            raise ValueError(
+                "ill-formated argument with odd number of quotes: %s" % pkg_req_str
+            )
+
+        if "<" in pkg_req_str or ">" in pkg_req_str:
+            if pkg_req_str[0] == '"' and pkg_req_str[-1] == '"':
+                # already double-quoted string
+                return pkg_req_str
+            else:
+                # need to double-quote string
+                if '"' in pkg_req_str:
+                    raise ValueError(
+                        "Cannot escape requirement string: %s" % pkg_req_str
+                    )
+                return '"%s"' % pkg_req_str
+        else:
+            # no dangerous char: no need to double-quote string
+            return pkg_req_str
+
+    # double-quote all args that need to be and return the result
+    return tuple(_dblquote_pkg_install_arg(a) for a in args)
 
 
 class _SessionQuit(Exception):
@@ -107,6 +138,11 @@ class Session:
         return {"_runner": self._runner}
 
     @property
+    def name(self) -> str:
+        """The name of this session."""
+        return self._runner.friendly_name
+
+    @property
     def env(self) -> dict:
         """A dictionary of environment variables to pass into all commands."""
         return self.virtualenv.env
@@ -131,16 +167,31 @@ class Session:
         return self._runner.func.python
 
     @property
-    def bin(self) -> Optional[str]:
-        """The bin directory for the virtualenv."""
-        return self.virtualenv.bin
+    def bin_paths(self) -> Optional[List[str]]:
+        """The bin directories for the virtualenv."""
+        return self.virtualenv.bin_paths
+
+    @property
+    def bin(self) -> str:
+        """The first bin directory for the virtualenv."""
+        paths = self.bin_paths
+        if paths is None:
+            raise ValueError("The environment does not have a bin directory.")
+        return paths[0]
+
+    def create_tmp(self) -> str:
+        """Create, and return, a temporary directory."""
+        tmpdir = os.path.join(self._runner.envdir, "tmp")
+        os.makedirs(tmpdir, exist_ok=True)
+        self.env["TMPDIR"] = tmpdir
+        return tmpdir
 
     @property
     def interactive(self) -> bool:
         """Returns True if Nox is being run in an interactive session or False otherwise."""
         return not self._runner.global_config.non_interactive and sys.stdin.isatty()
 
-    def chdir(self, dir: str) -> None:
+    def chdir(self, dir: Union[str, os.PathLike]) -> None:
         """Change the current working directory."""
         self.log("cd {}".format(dir))
         os.chdir(dir)
@@ -217,6 +268,34 @@ class Session:
 
         return self._run(*args, env=env, **kwargs)
 
+    def run_always(
+        self, *args: str, env: Mapping[str, str] = None, **kwargs: Any
+    ) -> Optional[Any]:
+        """Run a command **always**.
+
+        This is a variant of :meth:`run` that runs in all cases, including in
+        the presence of ``--install-only``.
+
+        :param env: A dictionary of environment variables to expose to the
+            command. By default, all environment variables are passed.
+        :type env: dict or None
+        :param bool silent: Silence command output, unless the command fails.
+            ``False`` by default.
+        :param success_codes: A list of return codes that are considered
+            successful. By default, only ``0`` is considered success.
+        :type success_codes: list, tuple, or None
+        :param external: If False (the default) then programs not in the
+            virtualenv path will cause a warning. If True, no warning will be
+            emitted. These warnings can be turned into errors using
+            ``--error-on-external-run``. This has no effect for sessions that
+            do not have a virtualenv.
+        :type external: bool
+        """
+        if not args:
+            raise ValueError("At least one argument required to run_always().")
+
+        return self._run(*args, env=env, **kwargs)
+
     def _run(self, *args: str, env: Mapping[str, str] = None, **kwargs: Any) -> Any:
         """Like run(), except that it runs even if --install-only is provided."""
         # Legacy support - run a function given.
@@ -243,9 +322,11 @@ class Session:
             kwargs["external"] = True
 
         # Run a shell command.
-        return nox.command.run(args, env=env, path=self.bin, **kwargs)
+        return nox.command.run(args, env=env, paths=self.bin_paths, **kwargs)
 
-    def conda_install(self, *args: str, **kwargs: Any) -> None:
+    def conda_install(
+        self, *args: str, auto_offline: bool = True, **kwargs: Any
+    ) -> None:
         """Install invokes `conda install`_ to install packages inside of the
         session's environment.
 
@@ -260,6 +341,10 @@ class Session:
             session.conda_install('--file', 'requirements.txt')
             session.conda_install('--file', 'requirements-dev.txt')
 
+        By default this method will detect when internet connection is not
+        available and will add the `--offline` flag automatically in that case.
+        To disable this behaviour, set `auto_offline=False`.
+
         To install the current package without clobbering conda-installed
         dependencies::
 
@@ -272,22 +357,37 @@ class Session:
         .. _conda install:
         """
         venv = self._runner.venv
-        if not isinstance(venv, CondaEnv):
+
+        prefix_args = ()  # type: Tuple[str, ...]
+        if isinstance(venv, CondaEnv):
+            prefix_args = ("--prefix", venv.location)
+        elif not isinstance(venv, PassthroughEnv):  # pragma: no cover
             raise ValueError(
                 "A session without a conda environment can not install dependencies from conda."
             )
+
         if not args:
             raise ValueError("At least one argument required to install().")
 
+        # Escape args that should be (conda-specific; pip install does not need this)
+        args = _dblquote_pkg_install_args(args)
+
         if "silent" not in kwargs:
             kwargs["silent"] = True
+
+        extraopts = ()  # type: Tuple[str, ...]
+        if auto_offline and venv.is_offline():
+            logger.warning(
+                "Automatically setting the `--offline` flag as conda repo seems unreachable."
+            )
+            extraopts = ("--offline",)
 
         self._run(
             "conda",
             "install",
             "--yes",
-            "--prefix",
-            venv.location,
+            *extraopts,
+            *prefix_args,
             *args,
             external="error",
             **kwargs
@@ -318,7 +418,9 @@ class Session:
 
         .. _pip: https://pip.readthedocs.org
         """
-        if not isinstance(self._runner.venv, (CondaEnv, VirtualEnv)):
+        if not isinstance(
+            self._runner.venv, (CondaEnv, VirtualEnv, PassthroughEnv)
+        ):  # pragma: no cover
             raise ValueError(
                 "A session without a virtualenv can not install dependencies."
             )
@@ -328,7 +430,7 @@ class Session:
         if "silent" not in kwargs:
             kwargs["silent"] = True
 
-        self._run("pip", "install", *args, external="error", **kwargs)
+        self._run("python", "-m", "pip", "install", *args, external="error", **kwargs)
 
     def notify(self, target: "Union[str, SessionRunner]") -> None:
         """Place the given session at the end of the queue.
@@ -388,33 +490,42 @@ class SessionRunner:
     def friendly_name(self) -> str:
         return self.signatures[0] if self.signatures else self.name
 
+    @property
+    def envdir(self) -> str:
+        return _normalize_path(self.global_config.envdir, self.friendly_name)
+
     def _create_venv(self) -> None:
-        if self.func.python is False:
-            self.venv = ProcessEnv()
+        backend = (
+            self.global_config.force_venv_backend
+            or self.func.venv_backend
+            or self.global_config.default_venv_backend
+        )
+
+        if backend == "none" or self.func.python is False:
+            self.venv = PassthroughEnv()
             return
 
-        path = _normalize_path(self.global_config.envdir, self.friendly_name)
         reuse_existing = (
             self.func.reuse_venv or self.global_config.reuse_existing_virtualenvs
         )
 
-        if not self.func.venv_backend or self.func.venv_backend == "virtualenv":
+        if backend is None or backend == "virtualenv":
             self.venv = VirtualEnv(
-                path,
+                self.envdir,
                 interpreter=self.func.python,  # type: ignore
                 reuse_existing=reuse_existing,
                 venv_params=self.func.venv_params,
             )
-        elif self.func.venv_backend == "conda":
+        elif backend == "conda":
             self.venv = CondaEnv(
-                path,
+                self.envdir,
                 interpreter=self.func.python,  # type: ignore
                 reuse_existing=reuse_existing,
                 venv_params=self.func.venv_params,
             )
-        elif self.func.venv_backend == "venv":
+        elif backend == "venv":
             self.venv = VirtualEnv(
-                path,
+                self.envdir,
                 interpreter=self.func.python,  # type: ignore
                 reuse_existing=reuse_existing,
                 venv=True,
@@ -423,7 +534,7 @@ class SessionRunner:
         else:
             raise ValueError(
                 "Expected venv_backend one of ('virtualenv', 'conda', 'venv'), but got '{}'.".format(
-                    self.func.venv_backend
+                    backend
                 )
             )
 
